Showing current tile
