+++ app/editor.py
@@ -2,6 +2,8 @@ from ._internal import *
 from .entities import *
 from .game import Game
 from .camera import Camera
+from .editor_camera import EditorCamera
+from .utils import *
 
 
 class Editor(Game):
@@ -56,7 +58,8 @@ class Editor(Game):
 		self.camera.update(
 			display=self.display,
 			top_left=pygame.Vector2((0, 0)),
-			static_player=True)
+			static_player=True,
+			current_obj=self.get_current_obj())
 
 		self.screen.blit(pygame.transform.scale(self.display, self.screen.get_size()), (0, 0))
 		pygame.display.update()
@@ -195,52 +198,7 @@ class Editor(Game):
 
 		self.update_current_obj()
 
-
-class EditorCamera(pygame.sprite.Sprite):
-	CAM_VEL = 20
-
-	def __init__(self, x, y, *groups) -> None:
-		super().__init__(*groups)
-		self.rect = pygame.Rect(x, y, SCREEN_WIDTH, SCREEN_HEIGHT)
-		self.scroll = pygame.Vector2((x, y))
-
-	def update(self, **kwargs) -> None:
-		"""
-		Call in game loop
-		"""
-		self.move()
-
-	def move(self) -> None:
-		keys = pygame.key.get_pressed()
-		dir_x = (keys[pygame.K_d] - keys[pygame.K_a]) * self.CAM_VEL
-		dir_y = (keys[pygame.K_s] - keys[pygame.K_w]) * self.CAM_VEL
-
-		# Normalize diagonal movement
-		if dir_x != 0 and dir_y != 0:
-			dir_x *= 0.7071
-			dir_y *= 0.7071
-
-		self.rect.x += dir_x
-		self.rect.y += dir_y
-		self.scroll.x += dir_x
-		self.scroll.y += dir_y
-		self.set_border()
-
-	def set_border(self) -> None:
-		self.rect.x = max(0, self.rect.x)
-		self.rect.y = max(0, self.rect.y)
-		self.scroll.x = max(0, self.scroll.x)
-		self.scroll.y = max(0, self.scroll.y)
-
-	def mpos_to_wpos(self, o_screen: pygame.Vector2) -> pygame.Vector2:
-		"""
-		Returns world position based on mouse position
-		"""
-		mpos = pygame.Vector2(pygame.mouse.get_pos())
-		ratio_x = SCREEN_WIDTH * CAM_SCALE / o_screen.x
-		ratio_y = SCREEN_HEIGHT * CAM_SCALE / o_screen.y
-		adjust = pygame.Vector2((mpos.x * ratio_x, mpos.y * ratio_y))
-
-		return pygame.Vector2((
-			(adjust.x + self.scroll.x) // RECT_WIDTH,
-			(adjust.y + self.scroll.y) // RECT_HEIGHT))
+	def get_current_obj(self) -> pygame.Surface:
+		current_obj = get_image([self.current_cat, self.current_obj], "1")
+		current_obj.set_alpha(150)
+		return current_obj
